implements animation system and example
 - removes obsolete dimension in favor of collision
 - adds new example main to showcase animation

+++ components/collision.go
@@ -1,8 +1,8 @@
 package components
 
 type Collision struct {
-	Width  int
-	Height int
+	Width  float64
+	Height float64
 }
 
 func (c *Collision) ID() string {
